[analyzer][NFC] Reorganize bstring.cpp tests
This change eliminates preprocessor-based suppression of test cases, and
moves special cases to separate test files.
